Control.Concurrent.Deque

Description

Chase-Lev work-stealing deques

This implementation derives directly from the pseudocode in the 2005 SPAA paper.

Documentation

data Deque a

A Chase-Lev circular work-stealing deque

A work-stealing deque is perhaps more properly referred to as a concurrent steque, where the dequeue-only steal side can be used concurrently, but the "local" side may that provides push and pop can only be used from a single owner thread.

Initialization

empty :: Arrayed a => IO (Deque a)

Create a new empty Deque.

>>> q :: Deque Int <- empty
>>> null q
True
>>> pop q
Nothing
>>> steal q
Nothing

singleton :: forall a. Arrayed a => a -> IO (Deque a)

Create a new Deque with one element in it.

>>> q <- singleton "hello"
>>> size q
(1,1)
>>> pop q
Just "hello"
>>> pop q
Nothing

fromList :: forall a. Arrayed a => [a] -> IO (Deque a)

Generate a work-stealing Deque of elements from a list.

>>> q <- fromList [1,2,3,4 :: Int]
>>> pop q
Just 1
>>> pop q
Just 2
>>> pop q
Just 3
>>> pop q
Just 4
>>> pop q
Nothing

>>> p <- fromList [1,2,3,4 :: Int]
>>> steal p
Just 4

fromListN :: forall a. Arrayed a => Int -> [a] -> IO (Deque a)

Generate a work-stealing Deque of elements from a list of known length.

Size

null :: Deque a -> IO Bool

null returns True if the Deque is definitely empty.

>>> q <- singleton (1 :: Int)
>>> null q
False
>>> pop q
Just 1
>>> null q
True

size :: Deque a -> IO (Int, Int)

Compute a lower and upper bound on the number of elements left in the Deque.

Under contention from stealing threads or when used by a stealing thread these numbers may well differ.

Local Operations

push :: Arrayed a => a -> Deque a -> IO ()

For a work-stealing queue push is the `local' push.

Thus only a single thread should perform this operation.

push and pop together act like a stack.

>>> q :: Deque String <- empty
>>> push "hello" q
>>> push "world" q
>>> pop q
Just "world"
>>> pop q
Just "hello"
>>> pop q
Nothing

push and steal together act like a queue.

>>> p :: Deque String <- empty
>>> push "hello" p
>>> push "world" p
>>> steal p
Just "hello"
>>> steal p
Just "world"
>>> steal p
Nothing

pushMany :: (Foldable f, Arrayed a) => f a -> Deque a -> IO ()

pop :: Arrayed a => Deque a -> IO (Maybe a)

Locally pop the Deque. This is not a thread safe operation, and should only be invoked on from the thread that "owns" the Deque.

Work-Stealing

steal :: Arrayed a => Deque a -> IO (Maybe a)

This is the work-stealing dequeue operation.

Multiple threads may concurrently attempt to steal from the same thread.

It may return Nothing during a lost race with another concurrent reader, so failing to read is not an indication that the Deque is empty!

However, at least one of the concurrently stealing threads will succeed.