Data.Vector.Array

Description

This module provides a choice of a best vector type for a given type that is as unboxed as possible.

Documentation

class (Vector (Arr a) a, Monoid (Arr a a)) => Arrayed a

Associated Types

type Arr a :: * -> *

type Array a = Arr a a

A vector of product-like data types that know how to store themselves in a Vector optimally, maximizing the level of unboxing provided, but not guaranteeing to unbox it all.

type MArray s a = Mutable (Arr a) s a

Internals

data V_Complex where

Constructors

V_Complex :: !Int -> !(Array a) -> !(Array a) -> V_Complex (Complex a)

data MV_Complex where

Constructors

MV_Complex :: !Int -> !(Mutable (Arr a) s a) -> !(Mutable (Arr a) s a) -> MV_Complex s (Complex a)

data V_Pair where

Constructors

V_Pair :: !Int -> !(Array a) -> !(Array b) -> V_Pair (a, b)

data MV_Pair where

Constructors

MV_Pair :: !Int -> !(Mutable (Arr a) s a) -> !(Mutable (Arr b) s b) -> MV_Pair s (a, b)