Data.Vector.Bit

Bits

newtype Bit

A simple newtype around a Bool

The principal use of this is that a Vector Bit is densely packed into individual bits rather than stored as one entry per Word8.

Constructors

Bit

Fields

getBit :: Bool

_Bit :: Iso' Bit Bool

Bit and Bool are isomorphic.

Bit Vectors with rank

data BitVector

A BitVector support for naïve O(1) rank.

Constructors

BitVector !Int !(Array Bit) !(Vector Int)

_BitVector :: Iso' BitVector (Array Bit)

O(n) embedding A BitVector is isomorphic to a vector of bits. It just carries extra information.

rank :: BitVector -> Int -> Int

O(1). rank i v counts the number of True bits up through and including the position i

null :: BitVector -> Bool

O(1). Is the BitVector empty?

size :: BitVector -> Int

O(1). Return the size of the BitVector.

singleton :: Bool -> BitVector

Construct a BitVector with a single element.

empty :: BitVector

The empty BitVector

Vectors of Bits

data family MVector s a

data family Vector a