Data.Vector.Bloom

Description

Hierarchical Bloom filters

Documentation

data Bloom

Constructors

Bloom !Int !Int !(Vector Word64)

Information

entries :: Bloom -> Int

Number of bits set

width :: Bloom -> Int

The number of bits in our Bloom filter. Always an integral multiple of 64.

Construction

bloom :: (Foldable f, Hashable a) => Int -> Int -> f a -> Bloom

bloom k m builds an m-bit wide Bloom filter that uses k hashes.

Modification

modify :: (forall s. MBloom s -> ST s ()) -> Bloom -> Bloom

Given an action on a mutable Bloom filter, modify this one.

insert :: Hashable a => a -> Bloom -> Bloom

Insert an element into a Bloom filter.

Testing

elem :: Hashable a => a -> Bloom -> Bool

Check if an element is a member of a Bloom filter.

This may return false positives, but never a false negative.

Combining Blooms

union :: Bloom -> Bloom -> Bloom

Compute the union of two Bloom filters.

intersection :: Bloom -> Bloom -> Bloom

Compute the intersection of two Bloom filters.

Freezing/Thawing

freeze :: PrimMonad m => MBloom (PrimState m) -> m Bloom

O(m)

thaw :: PrimMonad m => Bloom -> m (MBloom (PrimState m))

O(m)

unsafeFreeze :: PrimMonad m => MBloom (PrimState m) -> m Bloom

O(1)

unsafeThaw :: PrimMonad m => Bloom -> m (MBloom (PrimState m))

O(1)