Data.Vector.Map

Description

This module provides a functional variant on the Cache Oblivious Lookahead Array (COLA) by Bender et al. in "Cache-Oblivious Streaming B-Trees".

When used ephemerally, this Map has asymptotic performance equal to that of a B-Tree tuned to the parameters of your caches. However, no such parameter tuning is required.

Currently this Map is implemented in an insert-only fashion. Deletions are left to future work or to another derived structure in case they prove expensive.

Moreover, unlike the COLA from the paper, this version merely provides amortized complexity bounds as this permits us to provide a fully functional API. However, even those asymptotics are only guaranteed if you do not modify the "old" versions of the Map. If you do, then while correctness is preserved, the asymptotic analysis is inaccurate.

However, reading from "old" versions of the Map will not affect the asymptotic analysis.

Compared to Data.Map, this data structure currently consumes both more memory and more time, but it enables us to utilize contiguous storage.

Documentation

data Map k v

This Map is implemented as an insert-only Cache Oblivious Lookahead Array (COLA) with amortized complexity bounds that are equal to those of a B-Tree when it is used ephemerally, using Bloom filters to replace the fractional cascade.

Constructors

Map !Int (Maybe Bloom) !(Array k) !(Array v) !(Map k v)

Nil

empty :: Map k v

null :: Map k v -> Bool

singleton :: (Arrayed k, Arrayed v) => k -> v -> Map k v

lookup :: (Hashable k, Ord k, Arrayed k, Arrayed v) => k -> Map k v -> Maybe v

insert :: (Hashable k, Ord k, Arrayed k, Arrayed v) => k -> v -> Map k v -> Map k v

fromList :: (Hashable k, Ord k, Arrayed k, Arrayed v) => [(k, v)] -> Map k v

shape :: Map k v -> [Int]